Add remsh, cleanup modules, refactor
Added remote shell capabilities, refactored module setup with correct
require/provide lines as well as correct comments at the top and a
couchbase-pkg.el file so that it is a multifile 'package'.

Also added lots of http helpers and things to do the remote shell
stuff, but I can easily extend it to do a lot more now that we can
easily communicate with the nodes from the REST endpoint as well as
the erlang OTP distribution connection (albeit with 'inferior-erlang'
and not edts-shell or similar).