Latest changes to add completion for remsh + more
Added lots of functions.. cleaned up a lot of stuff, played aroudn
with and made examples of 'dash'. Made the display for ivy much
prettier with the optional description stuff being added now to
include what services are running on the box, its version string and
whether or not it has nodeEncryption enabled. Once it gets your
selection it does the rest and connect to the remote shell.

TODO: make it s/t we get completion in the remsh.. otherwise it's hard
to use. Alternatively I could make a very good doc lookup setup to
compensate but it would be easier if we just had the completion.