Latest fixes and changes
got rid of a lot of the annoying &rest business and maade the initial
one prompt for the host to connect to. TODO: make the default value
that pops up to prompt is already the default of 127.0.0.1:9000 for
development work. Lots of other fixes and changes as well as module
cleanup and renaming of completion module.