Create default host/port in remsh + start macros
Started creating macro-expansion stuff by working with how to handle
selecting the correct symbol from the buffer we are working on. That
is largely all I have done so far... As of now it seems like we can
either pipe it through `erlc` or load and use `epp` stuff manually
ourself and bring the data back and fourth. It seems like we are
"file-based" to some extent whether we want to or not. Is it better to
just shell out to the `erlc` command and then bringing back only the
section we want that was expanded. This is challening and would
require a bunch of heuristics to get right. Piping data through `epp`,
on the other hand, gives us much more information about the structure
of the code, and should make it easier to know what part we are
bringing back as the "expanded" macro. Basically, the epp way gives
some programatic advantages, while piping to erlc is probably just
simler to get started now..

TODO: figure that out.. lulz