WIP: shell completion stuff
Working on erlang shell completion. Unfortunately it has to be done
manually by scanning modules in elisp or getting it from the shell
itself (the way I am currently attempting to do it). The parsing is
what is happening now. Basically the process is:

 1. m().
 2. m(module_name1).
 3. m(module_name2).
 ...
 n. m(module_nameN).

... and parse and construct some hash maps that can be used to do
completion. This may seem insane but hte process is relatively simple
once the nasty parsing happens..